Move upload function into models
