+++ modelstore/model_store.py
@@ -13,16 +13,12 @@
 #    limitations under the License.
 import os
 import tarfile
-import uuid
 from dataclasses import dataclass
-from datetime import datetime
 
-import modelstore
 from modelstore.clouds.aws import BOTO_EXISTS, AWSStorage
 from modelstore.clouds.file_system import FileSystemStorage
 from modelstore.clouds.gcloud import GCLOUD_EXISTS, GoogleCloudStorage
 from modelstore.clouds.storage import CloudStorage
-from modelstore.meta import dependencies, revision, runtime
 from modelstore.models.managers import iter_libraries
 
 
@@ -78,40 +74,6 @@ class ModelStore:
         for library, manager in iter_libraries():
             object.__setattr__(self, library, manager)
 
-    def upload(self, domain: str, archive_path: str) -> dict:
-        """Upload an archive to cloud storage. This function returns
-        a dictionary of meta-data that is associated with this model,
-        including an id.
-        """
-        _validate_domain(domain)
-        model_id = str(uuid.uuid4())
-        # Warning! Mac OS translates ":" in paths to "/"
-        upload_time = datetime.now().strftime("%Y/%m/%d/%H:%M:%S")
-        location = self.storage.upload(domain, upload_time, archive_path)
-
-        meta_data = {
-            "model": {
-                "domain": domain,
-                "model_id": model_id,
-                "type": dependencies.extract_model_type(archive_path),
-            },
-            "storage": {"name": self.storage.get_name(), "location": location,},
-            "meta": {
-                "runtime": f"python:{runtime.get_python_version()}",
-                "user": runtime.get_user(),
-                "created": upload_time,
-                "dependencies": dependencies.extract_dependencies(archive_path),
-            },
-            "modelstore": modelstore.__version__,
-        }
-
-        git_meta = revision.git_meta()
-        if git_meta is not None:
-            meta_data["meta"]["git"] = git_meta
-
-        self.storage.set_meta_data(domain, model_id, meta_data)
-        return meta_data
-
     def list_domains(self) -> list:
         """ Returns a list of dicts, containing info about all
         of the domains """
@@ -131,15 +93,3 @@ class ModelStore:
             tar.extractall(local_path)
         os.remove(archive_path)
         return local_path
-
-
-def _validate_domain(domain: str):
-    if len(domain) == 0:
-        raise ValueError("Please provide a non-empty domain name.")
-    if domain in [
-        "versions",
-        "domains",
-        "modelstore",
-        "operatorai-model-store",
-    ]:
-        raise ValueError("Please use a different domain name.")
